divide event track proxy logic into proxy and handler
